Move icons out of mappings
