+++ src/browser/icons.ts
